# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/meal-plans": {
      "post": {
        "operationId": "MealPlansController_create",
        "summary": "Criar um novo plano alimentar",
        "description": "Cria um novo plano alimentar para um paciente com suas refeições",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plano alimentar criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlan"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MealPlansController_findAll",
        "summary": "Listar todos os planos alimentares",
        "description": "Retorna uma lista de todos os planos alimentares cadastrados",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de planos alimentares retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealPlan"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/meal-plans/patient/{patientId}": {
      "get": {
        "operationId": "MealPlansController_findByPatient",
        "summary": "Buscar planos por paciente",
        "description": "Retorna todos os planos alimentares de um paciente específico",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de planos alimentares retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealPlan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/meal-plans/{id}": {
      "get": {
        "operationId": "MealPlansController_findOne",
        "summary": "Buscar plano alimentar por ID",
        "description": "Retorna os detalhes de um plano alimentar específico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do plano alimentar",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plano alimentar encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlan"
                }
              }
            }
          },
          "404": {
            "description": "Plano alimentar não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "MealPlansController_update",
        "summary": "Atualizar plano alimentar",
        "description": "Atualiza as informações de um plano alimentar existente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do plano alimentar",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plano alimentar atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlan"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Plano alimentar não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MealPlansController_remove",
        "summary": "Remover plano alimentar",
        "description": "Remove um plano alimentar do sistema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do plano alimentar",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Plano alimentar removido com sucesso"
          },
          "404": {
            "description": "Plano alimentar não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/meal-plans/{id}/meals": {
      "post": {
        "operationId": "MealPlansController_addMeal",
        "summary": "Adicionar refeição",
        "description": "Adiciona uma nova refeição a um plano alimentar existente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do plano alimentar",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refeição adicionada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Plano alimentar não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MealPlansController_getMeals",
        "summary": "Listar refeições do plano",
        "description": "Retorna uma lista de refeições de um plano alimentar específico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do plano alimentar",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de refeições retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meal"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plano alimentar não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/meal-plans/search": {
      "get": {
        "operationId": "MealPlansController_search",
        "summary": "Buscar planos de alimentação",
        "description": "Busca planos de alimentação por nome",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Termo de busca (nome do plano)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de planos encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealPlan"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/search": {
      "get": {
        "operationId": "SearchController_search",
        "summary": "Busca global por pacientes e planos alimentares",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Termo de busca (mínimo 3 caracteres)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna uma lista combinada de pacientes e planos alimentares"
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartNutri API",
    "description": "API para gestão completa de nutrição",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do paciente"
          },
          "name": {
            "type": "string",
            "description": "Nome do paciente"
          },
          "email": {
            "type": "string",
            "description": "Email do paciente"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do paciente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do paciente"
          },
          "address": {
            "type": "string",
            "description": "Endereço do paciente"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "cpf",
          "phone",
          "address",
          "nutritionistId"
        ]
      },
...
      "CreateMealFoodDto": {
        "type": "object",
        "properties": {
          "foodId": {
            "type": "string",
            "description": "ID do alimento",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "amount": {
            "type": "number",
            "description": "Quantidade do alimento",
            "example": 100
          },
          "unit": {
            "type": "string",
            "description": "Unidade de medida",
            "example": "g"
          }
        },
        "required": [
          "foodId",
          "amount",
          "unit"
        ]
      },
      "CreateMealDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da refeição",
            "example": "Café da manhã"
          },
          "time": {
            "type": "string",
            "description": "Horário da refeição",
            "example": "08:00"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre a refeição"
          },
          "mealFoods": {
            "description": "Lista de alimentos da refeição",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMealFoodDto"
            }
          }
        },
        "required": [
          "name",
          "time",
          "mealFoods"
        ]
      },
      "CreateMealPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do plano alimentar",
            "example": "Plano Alimentar Semanal"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre o plano alimentar",
            "example": "Plano alimentar para perda de peso"
          },
          "startDate": {
            "type": "string",
            "description": "Data de início do plano",
            "example": "2024-03-20"
          },
          "endDate": {
            "type": "string",
            "description": "Data de término do plano",
            "example": "2024-03-27"
          },
          "patientId": {
            "type": "string",
            "description": "ID do paciente",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista responsável",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "meals": {
            "description": "Lista de refeições do plano",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMealDto"
            }
          }
        },
        "required": [
          "name",
          "startDate",
          "endDate",
          "patientId",
          "nutritionistId",
          "meals"
        ]
      },
      "Nutritionist": {
        "type": "object",
        "properties": {}
      },
      "Meal": {
        "type": "object",
        "properties": {
          "totalCalories": {
            "type": "number",
            "example": 500.25
          },
          "totalProtein": {
            "type": "number",
            "example": 30.5
          },
          "totalCarbs": {
            "type": "number",
            "example": 60.75
          },
          "totalFat": {
            "type": "number",
            "example": 20.3
          }
        },
        "required": [
          "totalCalories",
          "totalProtein",
          "totalCarbs",
          "totalFat"
        ]
      },
      "MealPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "Plano Alimentar Semanal"
          },
          "notes": {
            "type": "string",
            "example": "Plano alimentar para perda de peso"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-27"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "patientId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nutritionist": {
            "$ref": "#/components/schemas/Nutritionist"
          },
          "nutritionistId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20T10:00:00Z"
          },
          "dailyCalories": {
            "type": "number",
            "example": 2000.5
          },
          "dailyProtein": {
            "type": "number",
            "example": 150.75
          },
          "dailyCarbs": {
            "type": "number",
            "example": 250.3
          },
          "dailyFat": {
            "type": "number",
            "example": 80.2
          }
        },
        "required": [
          "id",
          "name",
          "notes",
          "startDate",
          "endDate",
          "patient",
          "patientId",
          "nutritionist",
          "nutritionistId",
          "meals",
          "createdAt",
          "updatedAt",
          "dailyCalories",
          "dailyProtein",
          "dailyCarbs",
          "dailyFat"
        ]
      },
      "UpdateMealPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do plano alimentar",
            "example": "Plano Semanal"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "Data de início do plano",
            "example": "2025-04-05"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "description": "Data de término do plano",
            "example": "2025-04-12"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre o plano"
          },
          "meals": {
            "description": "Lista de refeições do plano",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateMealDto"
            }
          }
        }
      },
...
    }
  }
}